requirements.txt
psutil
python-dotenv
pytz
telethon